test: harden Batch conformance coverage
Make open-ledger Batch handling and exact TER assertions match rippled, verify closed inner metadata and signed paths, and keep the Batch fixture fail-closed and test-only.\n\nFixes #1499
