+++ internal/testing/batch/open_ledger_test.go
@@ -1,0 +1,927 @@
+package batch
+
+import (
+	"fmt"
+	"testing"
+
+	jtx "github.com/LeJamon/go-xrpl/internal/testing"
+	"github.com/LeJamon/go-xrpl/internal/tx"
+	accounttx "github.com/LeJamon/go-xrpl/internal/tx/account"
+	batchtx "github.com/LeJamon/go-xrpl/internal/tx/batch"
+	"github.com/LeJamon/go-xrpl/internal/tx/check"
+	"github.com/LeJamon/go-xrpl/internal/tx/payment"
+	"github.com/LeJamon/go-xrpl/internal/txq"
+	"github.com/stretchr/testify/require"
+)
+
+func TestTickets(t *testing.T) {
+	t.Run("tickets outer", func(t *testing.T) {
+		// Outer batch uses a ticket; inner transactions use regular sequences.
+		// Reference: rippled Batch_test.cpp testTickets() - "tickets outer"
+		env := newBatchEnv(t)
+
+		alice := jtx.NewAccount("alice")
+		bob := jtx.NewAccount("bob")
+		env.FundAmount(alice, uint64(jtx.XRP(10000)))
+		env.FundAmount(bob, uint64(jtx.XRP(10000)))
+		env.Close()
+
+		// Create 10 tickets for alice
+		aliceTicketSeq := env.CreateTickets(alice, 10)
+		env.Close()
+
+		aliceSeq := env.Seq(alice)
+		preAlice := env.Balance(alice)
+		preBob := env.Balance(bob)
+
+		// Submit batch with outer using ticket, inner using sequences
+		batchFee := CalcBatchFeeFromEnv(env, 0, 2)
+		batch := NewBatchBuilderWithTicket(alice, aliceTicketSeq, batchFee, batchtx.BatchFlagAllOrNothing).
+			AddInnerTx(MakeInnerPaymentXRP(alice, bob, 1, aliceSeq+0)).
+			AddInnerTx(MakeInnerPaymentXRP(alice, bob, 2, aliceSeq+1)).
+			MustBuild()
+
+		result := env.Submit(batch)
+		jtx.RequireTxSuccess(t, result)
+		env.Close()
+
+		// Verify owner count: started with 10 tickets, consumed 1 (outer ticket) = 9
+		require.Equal(t, uint32(9), env.OwnerCount(alice), "alice should have 9 owner objects")
+		require.Equal(t, uint32(9), env.TicketCount(alice), "alice should have 9 tickets remaining")
+
+		// Alice's sequence advances by 2 (inner txns use sequences)
+		jtx.RequireSequence(t, env, alice, aliceSeq+2)
+
+		// Alice pays XRP(3) + batchFee; Bob receives XRP(3)
+		jtx.RequireBalance(t, env, alice, preAlice-uint64(jtx.XRP(3))-batchFee)
+		jtx.RequireBalance(t, env, bob, preBob+uint64(jtx.XRP(3)))
+	})
+
+	t.Run("tickets inner", func(t *testing.T) {
+		// Outer batch uses regular sequence; inner transactions use tickets.
+		// Reference: rippled Batch_test.cpp testTickets() - "tickets inner"
+		env := newBatchEnv(t)
+
+		alice := jtx.NewAccount("alice")
+		bob := jtx.NewAccount("bob")
+		env.FundAmount(alice, uint64(jtx.XRP(10000)))
+		env.FundAmount(bob, uint64(jtx.XRP(10000)))
+		env.Close()
+
+		// Create 10 tickets for alice
+		aliceTicketSeq := env.CreateTickets(alice, 10)
+		env.Close()
+
+		aliceSeq := env.Seq(alice)
+		preAlice := env.Balance(alice)
+		preBob := env.Balance(bob)
+
+		// Submit batch with outer using sequence, inner using tickets
+		batchFee := CalcBatchFeeFromEnv(env, 0, 2)
+		batch := NewBatchBuilder(alice, aliceSeq, batchFee, batchtx.BatchFlagAllOrNothing).
+			AddInnerTx(MakeInnerPaymentXRPWithTicket(alice, bob, 1, aliceTicketSeq)).
+			AddInnerTx(MakeInnerPaymentXRPWithTicket(alice, bob, 2, aliceTicketSeq+1)).
+			MustBuild()
+
+		result := env.Submit(batch)
+		jtx.RequireTxSuccess(t, result)
+		env.Close()
+
+		// Verify owner count: started with 10 tickets, consumed 2 (inner tickets) = 8
+		require.Equal(t, uint32(8), env.OwnerCount(alice), "alice should have 8 owner objects")
+		require.Equal(t, uint32(8), env.TicketCount(alice), "alice should have 8 tickets remaining")
+
+		// Alice's sequence advances by 1 (only outer seq increment, inner use tickets)
+		jtx.RequireSequence(t, env, alice, aliceSeq+1)
+
+		// Alice pays XRP(3) + batchFee; Bob receives XRP(3)
+		jtx.RequireBalance(t, env, alice, preAlice-uint64(jtx.XRP(3))-batchFee)
+		jtx.RequireBalance(t, env, bob, preBob+uint64(jtx.XRP(3)))
+	})
+
+	t.Run("tickets outer inner", func(t *testing.T) {
+		// Outer batch uses a ticket; one inner tx uses a ticket, the other uses a sequence.
+		// Reference: rippled Batch_test.cpp testTickets() - "tickets outer inner"
+		env := newBatchEnv(t)
+
+		alice := jtx.NewAccount("alice")
+		bob := jtx.NewAccount("bob")
+		env.FundAmount(alice, uint64(jtx.XRP(10000)))
+		env.FundAmount(bob, uint64(jtx.XRP(10000)))
+		env.Close()
+
+		// Create 10 tickets for alice
+		aliceTicketSeq := env.CreateTickets(alice, 10)
+		env.Close()
+
+		aliceSeq := env.Seq(alice)
+		preAlice := env.Balance(alice)
+		preBob := env.Balance(bob)
+
+		// Submit batch:
+		// - outer uses ticket aliceTicketSeq
+		// - inner[0] uses ticket aliceTicketSeq+1
+		// - inner[1] uses sequence aliceSeq
+		batchFee := CalcBatchFeeFromEnv(env, 0, 2)
+		batch := NewBatchBuilderWithTicket(alice, aliceTicketSeq, batchFee, batchtx.BatchFlagAllOrNothing).
+			AddInnerTx(MakeInnerPaymentXRPWithTicket(alice, bob, 1, aliceTicketSeq+1)).
+			AddInnerTx(MakeInnerPaymentXRP(alice, bob, 2, aliceSeq)).
+			MustBuild()
+
+		result := env.Submit(batch)
+		jtx.RequireTxSuccess(t, result)
+		env.Close()
+
+		// Verify owner count: started with 10 tickets, consumed 2 (outer + inner[0]) = 8
+		require.Equal(t, uint32(8), env.OwnerCount(alice), "alice should have 8 owner objects")
+		require.Equal(t, uint32(8), env.TicketCount(alice), "alice should have 8 tickets remaining")
+
+		// Alice's sequence advances by 1 (only inner[1] uses a sequence)
+		jtx.RequireSequence(t, env, alice, aliceSeq+1)
+
+		// Alice pays XRP(3) + batchFee; Bob receives XRP(3)
+		jtx.RequireBalance(t, env, alice, preAlice-uint64(jtx.XRP(3))-batchFee)
+		jtx.RequireBalance(t, env, bob, preBob+uint64(jtx.XRP(3)))
+	})
+}
+
+func TestTicketsOpenLedger(t *testing.T) {
+	// Reference: rippled Batch_test.cpp testTicketsOpenLedger()
+	// Tests that batch transactions using tickets interact correctly with
+	// standalone transactions using tickets from the same set.
+
+	t.Run("before batch txn with same ticket", func(t *testing.T) {
+		// Reference: rippled testTicketsOpenLedger() "Before Batch Txn w/ same ticket"
+		// The batch is applied first (canonical order), consuming the ticket
+		// used by the inner tx. The noop that also uses that ticket is
+		// overwritten.
+		env := newBatchEnv(t)
+		env.EnableOpenLedgerReplay()
+
+		alice := jtx.NewAccount("alice")
+		bob := jtx.NewAccount("bob")
+		env.Fund(alice, bob)
+		env.Close()
+
+		aliceTicketSeq := env.CreateTickets(alice, 10)
+		env.Close()
+
+		aliceSeq := env.Seq(alice)
+
+		// AccountSet Txn using ticket+1
+		noopTxn := accounttx.NewAccountSet(alice.Address)
+		noopTxn.Fee = fmt.Sprintf("%d", env.BaseFee())
+		noopTxn.SigningPubKey = alice.PublicKeyHex()
+		zero := uint32(0)
+		noopTxn.Sequence = &zero
+		ticketSeq1 := aliceTicketSeq + 1
+		noopTxn.TicketSequence = &ticketSeq1
+		result := env.Submit(noopTxn)
+		jtx.RequireTxSuccess(t, result)
+
+		// Batch Txn using ticket for outer, ticket+1 for inner payment
+		batchFee := CalcBatchFeeFromEnv(env, 0, 2)
+		batch := NewBatchBuilderWithTicket(alice, aliceTicketSeq, batchFee, batchtx.BatchFlagAllOrNothing).
+			AddInnerTx(MakeInnerPaymentXRPWithTicket(alice, bob, 1, aliceTicketSeq+1)).
+			AddInnerTx(MakeInnerPaymentXRP(alice, bob, 2, aliceSeq)).
+			MustBuild()
+		result = env.Submit(batch)
+		jtx.RequireTxSuccess(t, result)
+		env.Close()
+
+		// After close: batch succeeds. The inner tx consumed ticket+1, so
+		// the standalone noop that used ticket+1 fails during replay
+		// (ticket already consumed). alice seq should advance.
+		env.Close()
+
+		// Verify final state: alice consumed aliceSeq (inner payment #2),
+		// ticket (batch outer), ticket+1 (inner payment #1)
+		require.Equal(t, aliceSeq+1, env.Seq(alice))
+	})
+
+	t.Run("after batch txn with same ticket", func(t *testing.T) {
+		// Reference: rippled testTicketsOpenLedger() "After Batch Txn w/ same ticket"
+		env := newBatchEnv(t)
+		env.EnableOpenLedgerReplay()
+
+		alice := jtx.NewAccount("alice")
+		bob := jtx.NewAccount("bob")
+		env.Fund(alice, bob)
+		env.Close()
+
+		aliceTicketSeq := env.CreateTickets(alice, 10)
+		env.Close()
+
+		aliceSeq := env.Seq(alice)
+
+		// Batch Txn first
+		batchFee := CalcBatchFeeFromEnv(env, 0, 2)
+		batch := NewBatchBuilderWithTicket(alice, aliceTicketSeq, batchFee, batchtx.BatchFlagAllOrNothing).
+			AddInnerTx(MakeInnerPaymentXRPWithTicket(alice, bob, 1, aliceTicketSeq+1)).
+			AddInnerTx(MakeInnerPaymentXRP(alice, bob, 2, aliceSeq)).
+			MustBuild()
+		result := env.Submit(batch)
+		jtx.RequireTxSuccess(t, result)
+
+		// AccountSet Txn using ticket+1 (already consumed by batch inner)
+		noopTxn := accounttx.NewAccountSet(alice.Address)
+		noopTxn.Fee = fmt.Sprintf("%d", env.BaseFee())
+		noopTxn.SigningPubKey = alice.PublicKeyHex()
+		zero := uint32(0)
+		noopTxn.Sequence = &zero
+		ticketSeq1 := aliceTicketSeq + 1
+		noopTxn.TicketSequence = &ticketSeq1
+		noopResult := env.Submit(noopTxn)
+		jtx.RequireTxSuccess(t, noopResult)
+		env.Close()
+
+		env.Close()
+
+		// Verify final state
+		require.Equal(t, aliceSeq+1, env.Seq(alice))
+	})
+}
+
+func TestBatchTxQueue(t *testing.T) {
+	t.Run("outer batch txns count towards queue size", func(t *testing.T) {
+		// Reference: rippled Batch_test.cpp testBatchTxQueue() first sub-test
+		// "only outer batch transactions are counter towards the queue size"
+		cfg := makeSmallQueueConfig(2)
+		env := jtx.NewTestEnvWithTxQ(t, cfg)
+		env.EnableFeatureNow("Batch")
+
+		alice := jtx.NewAccount("alice")
+		bob := jtx.NewAccount("bob")
+		carol := jtx.NewAccount("carol")
+
+		// Fund across several ledgers so the TxQ metrics stay restricted.
+		// noripple funds do not enable DefaultRipple (1 tx per account instead of 2).
+		env.FundAmountNoRipple(alice, uint64(jtx.XRP(10000)))
+		env.FundAmountNoRipple(bob, uint64(jtx.XRP(10000)))
+		env.Close()
+		env.FundAmountNoRipple(carol, uint64(jtx.XRP(10000)))
+		env.Close()
+
+		// Fill the ledger: 3 noops above the threshold of 2.
+		env.Noop(alice)
+		env.Noop(alice)
+		env.Noop(alice)
+		checkMetrics(t, env, 0, nil, 3, 2)
+
+		// Carol's noop gets queued because fee escalation requires more than base fee.
+		result := env.Submit(makeNoopWithFee(carol, env.BaseFee()))
+		jtx.RequireTxFail(t, result, "terQUEUED")
+		checkMetrics(t, env, 1, nil, 3, 2)
+
+		aliceSeq := env.Seq(alice)
+		bobSeq := env.Seq(bob)
+		batchFee := CalcBatchFeeFromEnv(env, 1, 2)
+
+		// Queue Batch: regular batch fee is too low to bypass escalation.
+		batch := NewBatchBuilder(alice, aliceSeq, batchFee, batchtx.BatchFlagAllOrNothing).
+			AddInnerTx(MakeInnerPaymentXRP(alice, bob, 10, aliceSeq+1)).
+			AddInnerTx(MakeInnerPaymentXRP(bob, alice, 5, bobSeq)).
+			AddSigner(bob).
+			MustBuild()
+		result = env.Submit(batch)
+		jtx.RequireTxFail(t, result, "terQUEUED")
+		checkMetrics(t, env, 2, nil, 3, 2)
+
+		// Replace Queued Batch with open ledger fee.
+		olFee := env.OpenLedgerFee(batchFee)
+		batch2 := NewBatchBuilder(alice, aliceSeq, olFee, batchtx.BatchFlagAllOrNothing).
+			AddInnerTx(MakeInnerPaymentXRP(alice, bob, 10, aliceSeq+1)).
+			AddInnerTx(MakeInnerPaymentXRP(bob, alice, 5, bobSeq)).
+			AddSigner(bob).
+			MustBuild()
+		result = env.Submit(batch2)
+		jtx.RequireTxSuccess(t, result)
+		env.Close()
+
+		// After close: queue drained (carol's noop applied in new ledger),
+		// maxSize = txnsExpected * ledgersInQueue.
+		// Closed ledger had: 3 noops + 1 batch outer + 2 inner = 6 txns.
+		// With NormalConsensusIncreasePercent=0, txnsExpected = 6.
+		// maxSize = 6 * 2 = 12. txInLedger = 1 (carol's noop from queue).
+		maxSize := uint64(12)
+		checkMetrics(t, env, 0, &maxSize, 1, 6)
+		closed := env.LastClosedLedger()
+		require.Equal(t, uint32(6), closed.TxCount())
+		root, err := closed.TxMapHash()
+		require.NoError(t, err)
+		require.NotEqual(t, [32]byte{}, root)
+		require.Equal(t, root, closed.Header().TxHash)
+	})
+
+	t.Run("inner batch txns count towards ledger tx count", func(t *testing.T) {
+		// Reference: rippled Batch_test.cpp testBatchTxQueue() second sub-test
+		// "inner batch transactions are counter towards the ledger tx count"
+		cfg := makeSmallQueueConfig(2)
+		env := jtx.NewTestEnvWithTxQ(t, cfg)
+		env.EnableFeatureNow("Batch")
+
+		alice := jtx.NewAccount("alice")
+		bob := jtx.NewAccount("bob")
+		carol := jtx.NewAccount("carol")
+
+		// Fund across several ledgers so the TxQ metrics stay restricted.
+		env.FundAmountNoRipple(alice, uint64(jtx.XRP(10000)))
+		env.FundAmountNoRipple(bob, uint64(jtx.XRP(10000)))
+		env.Close()
+		env.FundAmountNoRipple(carol, uint64(jtx.XRP(10000)))
+		env.Close()
+
+		// Fill the ledger leaving room for 1 more transaction at base fee.
+		env.Noop(alice)
+		env.Noop(alice)
+		checkMetrics(t, env, 0, nil, 2, 2)
+
+		aliceSeq := env.Seq(alice)
+		bobSeq := env.Seq(bob)
+		batchFee := CalcBatchFeeFromEnv(env, 1, 2)
+
+		// Batch Successful: at txInLedger=2 (equal to threshold), batch gets in.
+		batch := NewBatchBuilder(alice, aliceSeq, batchFee, batchtx.BatchFlagAllOrNothing).
+			AddInnerTx(MakeInnerPaymentXRP(alice, bob, 10, aliceSeq+1)).
+			AddInnerTx(MakeInnerPaymentXRP(bob, alice, 5, bobSeq)).
+			AddSigner(bob).
+			MustBuild()
+		result := env.Submit(batch)
+		jtx.RequireTxSuccess(t, result)
+		// txInLedger = 3 (2 noops + 1 batch outer).
+		// The batch counts as 1 for open ledger txInLedger.
+		checkMetrics(t, env, 0, nil, 3, 2)
+
+		// Carol's noop gets queued because txInLedger=3 > txnsExpected=2.
+		result = env.Submit(makeNoopWithFee(carol, env.BaseFee()))
+		jtx.RequireTxFail(t, result, "terQUEUED")
+		checkMetrics(t, env, 1, nil, 3, 2)
+	})
+}
+
+// makeSmallQueueConfig creates a TxQ config matching rippled's Batch_test
+// makeSmallQueueConfig({{"minimum_txn_in_ledger_standalone", minTxn}}).
+// Reference: rippled Batch_test.cpp makeSmallQueueConfig()
+func makeSmallQueueConfig(minTxnStandalone uint32) txq.Config {
+	return txq.Config{
+		LedgersInQueue:                 2,
+		QueueSizeMin:                   2,
+		RetrySequencePercent:           25,
+		MinimumEscalationMultiplier:    txq.BaseLevel * 500, // 128000
+		MinimumTxnInLedger:             32,
+		MinimumTxnInLedgerStandalone:   minTxnStandalone,
+		TargetTxnInLedger:              256,
+		MaximumTxnInLedger:             0,
+		NormalConsensusIncreasePercent: 0,
+		SlowConsensusDecreasePercent:   50,
+		MaximumTxnPerAccount:           10,
+		MinimumLastLedgerBuffer:        2,
+		Standalone:                     true,
+	}
+}
+
+// makeNoopWithFee creates an AccountSet noop with a specific fee.
+// Unlike env.Noop() which auto-fills and submits, this returns the raw tx.
+func makeNoopWithFee(acc *jtx.Account, fee uint64) *accounttx.AccountSet {
+	as := accounttx.NewAccountSet(acc.Address)
+	as.Fee = fmt.Sprintf("%d", fee)
+	return as
+}
+
+// checkMetrics asserts TxQ metrics match expected values.
+// maxSize nil means skip that assertion (matches rippled's std::nullopt).
+// Reference: rippled test/jtx/TestHelpers.h checkMetrics()
+func checkMetrics(t *testing.T, env *jtx.TestEnv, expectedQueueSize uint64, expectedMaxSize *uint64, expectedTxInLedger uint64, expectedTxPerLedger uint64) {
+	t.Helper()
+	metrics := env.TxQMetrics()
+
+	require.Equal(t, expectedQueueSize, metrics.TxCount,
+		"checkMetrics: txCount (queue size) mismatch")
+
+	if expectedMaxSize != nil {
+		require.NotNil(t, metrics.TxQMaxSize, "checkMetrics: maxSize should not be nil")
+		require.Equal(t, *expectedMaxSize, *metrics.TxQMaxSize,
+			"checkMetrics: txQMaxSize mismatch")
+	}
+
+	require.Equal(t, expectedTxInLedger, metrics.TxInLedger,
+		"checkMetrics: txInLedger mismatch")
+
+	require.Equal(t, expectedTxPerLedger, metrics.TxPerLedger,
+		"checkMetrics: txPerLedger mismatch")
+}
+
+func TestSequenceOpenLedger(t *testing.T) {
+	// Reference: rippled Batch_test.cpp testSequenceOpenLedger()
+	// Tests interactions between batch inner transactions advancing sequences
+	// and standalone transactions with future sequences or the same sequences.
+
+	t.Run("before batch txn with retry following ledger", func(t *testing.T) {
+		// Reference: rippled testSequenceOpenLedger() "Before Batch Txn w/ retry following ledger"
+		// A noop at aliceSeq+2 gets terPRE_SEQ. Then a batch with carol as
+		// outer submitter and alice as inner signer advances alice's seq.
+		// After close: batch succeeds, noop retried in next ledger.
+		env := newBatchEnv(t)
+		env.EnableOpenLedgerReplay()
+
+		alice := jtx.NewAccount("alice")
+		bob := jtx.NewAccount("bob")
+		carol := jtx.NewAccount("carol")
+		env.Fund(alice, bob, carol)
+		env.Close()
+
+		aliceSeq := env.Seq(alice)
+		carolSeq := env.Seq(carol)
+		preAlice := env.Balance(alice)
+		preBob := env.Balance(bob)
+
+		// AccountSet Txn at aliceSeq+2 -> terPRE_SEQ (future sequence)
+		noopTxn := accounttx.NewAccountSet(alice.Address)
+		noopTxn.Fee = fmt.Sprintf("%d", env.BaseFee())
+		noopTxn.SigningPubKey = alice.PublicKeyHex()
+		futureSeq := aliceSeq + 2
+		noopTxn.Sequence = &futureSeq
+		result := env.Submit(noopTxn)
+		jtx.RequireTxFail(t, result, "terPRE_SEQ")
+
+		// Batch Txn: carol outer, alice inner signer
+		batchFee := CalcBatchFeeFromEnv(env, 1, 2)
+		batch := NewBatchBuilder(carol, carolSeq, batchFee, batchtx.BatchFlagAllOrNothing).
+			AddInnerTx(MakeInnerPaymentXRP(alice, bob, 1, aliceSeq)).
+			AddInnerTx(MakeInnerPaymentXRP(alice, bob, 2, aliceSeq+1)).
+			AddSigner(alice).
+			MustBuild()
+		result = env.Submit(batch)
+		jtx.RequireTxSuccess(t, result)
+		env.Close()
+
+		// After first close: batch applied (alice seq -> aliceSeq+2).
+		// Noop at aliceSeq+2 may or may not be in first closed ledger
+		// (depends on canonical ordering and replay behavior).
+		// Close again to ensure the noop is retried if held.
+		env.Close()
+
+		// Final state verification:
+		// - alice's seq should be aliceSeq+3 (aliceSeq consumed by inner pay#1,
+		//   aliceSeq+1 by inner pay#2, aliceSeq+2 by noop)
+		require.Equal(t, aliceSeq+3, env.Seq(alice),
+			"alice seq should have advanced by 3 (2 inner payments + 1 noop)")
+
+		// Carol consumed carolSeq for the batch outer
+		require.Equal(t, carolSeq+1, env.Seq(carol),
+			"carol seq should have advanced by 1 (batch outer)")
+
+		// alice paid XRP(1) + XRP(2) = XRP(3) to bob, plus baseFee for noop
+		jtx.RequireBalance(t, env, alice, preAlice-uint64(jtx.XRP(3))-env.BaseFee())
+		jtx.RequireBalance(t, env, bob, preBob+uint64(jtx.XRP(3)))
+	})
+
+	t.Run("before batch txn with same sequence", func(t *testing.T) {
+		// Reference: rippled testSequenceOpenLedger() "Before Batch Txn w/ same sequence"
+		// A noop at aliceSeq+1 gets terPRE_SEQ. Then a batch with alice as
+		// outer submitter has inner payments consuming aliceSeq+1 and aliceSeq+2.
+		// After close: batch wins (canonical order), noop overwritten.
+		env := newBatchEnv(t)
+		env.EnableOpenLedgerReplay()
+
+		alice := jtx.NewAccount("alice")
+		bob := jtx.NewAccount("bob")
+		env.Fund(alice, bob)
+		env.Close()
+
+		aliceSeq := env.Seq(alice)
+		preAlice := env.Balance(alice)
+		preBob := env.Balance(bob)
+
+		// AccountSet Txn at aliceSeq+1 -> terPRE_SEQ
+		noopTxn := accounttx.NewAccountSet(alice.Address)
+		noopTxn.Fee = fmt.Sprintf("%d", env.BaseFee())
+		noopTxn.SigningPubKey = alice.PublicKeyHex()
+		futureSeq := aliceSeq + 1
+		noopTxn.Sequence = &futureSeq
+		result := env.Submit(noopTxn)
+		jtx.RequireTxFail(t, result, "terPRE_SEQ")
+
+		// Batch Txn: alice outer
+		batchFee := CalcBatchFeeFromEnv(env, 0, 2)
+		batch := NewBatchBuilder(alice, aliceSeq, batchFee, batchtx.BatchFlagAllOrNothing).
+			AddInnerTx(MakeInnerPaymentXRP(alice, bob, 1, aliceSeq+1)).
+			AddInnerTx(MakeInnerPaymentXRP(alice, bob, 2, aliceSeq+2)).
+			MustBuild()
+		result = env.Submit(batch)
+		jtx.RequireTxSuccess(t, result)
+		env.Close()
+
+		// After first close: batch applied. The noop at aliceSeq+1 is
+		// overwritten by the batch's inner payment at the same sequence.
+		// Close again to flush held transactions.
+		env.Close()
+
+		// Final state: batch consumed aliceSeq (outer), aliceSeq+1 (inner#1),
+		// aliceSeq+2 (inner#2). The noop at aliceSeq+1 failed (sequence
+		// already consumed by inner payment). alice seq = aliceSeq+3.
+		require.Equal(t, aliceSeq+3, env.Seq(alice),
+			"alice seq should have advanced by 3 (outer + 2 inner payments)")
+
+		// alice paid XRP(1) + XRP(2) + batchFee to bob
+		jtx.RequireBalance(t, env, alice, preAlice-uint64(jtx.XRP(3))-batchFee)
+		jtx.RequireBalance(t, env, bob, preBob+uint64(jtx.XRP(3)))
+	})
+
+	t.Run("after batch txn with same sequence", func(t *testing.T) {
+		// Reference: rippled testSequenceOpenLedger() "After Batch Txn w/ same sequence"
+		// Batch submitted first, then noop at aliceSeq+1.
+		// After close: batch wins (applied first), noop at same seq overwritten.
+		env := newBatchEnv(t)
+		env.EnableOpenLedgerReplay()
+
+		alice := jtx.NewAccount("alice")
+		bob := jtx.NewAccount("bob")
+		env.Fund(alice, bob)
+		env.Close()
+
+		aliceSeq := env.Seq(alice)
+		preAlice := env.Balance(alice)
+		preBob := env.Balance(bob)
+
+		// Batch Txn: alice outer
+		batchFee := CalcBatchFeeFromEnv(env, 0, 2)
+		batch := NewBatchBuilder(alice, aliceSeq, batchFee, batchtx.BatchFlagAllOrNothing).
+			AddInnerTx(MakeInnerPaymentXRP(alice, bob, 1, aliceSeq+1)).
+			AddInnerTx(MakeInnerPaymentXRP(alice, bob, 2, aliceSeq+2)).
+			MustBuild()
+		result := env.Submit(batch)
+		jtx.RequireTxSuccess(t, result)
+
+		// AccountSet Txn at aliceSeq+1 (same as inner payment #1's seq).
+		noopTxn := accounttx.NewAccountSet(alice.Address)
+		noopTxn.Fee = fmt.Sprintf("%d", env.BaseFee())
+		noopTxn.SigningPubKey = alice.PublicKeyHex()
+		sameSeq := aliceSeq + 1
+		noopTxn.Sequence = &sameSeq
+		noopResult := env.Submit(noopTxn)
+		jtx.RequireTxSuccess(t, noopResult)
+		env.Close()
+
+		// After close: batch consumed aliceSeq (outer), aliceSeq+1 (inner#1),
+		// aliceSeq+2 (inner#2). The noop at aliceSeq+1 was overwritten.
+		env.Close()
+
+		// Final state: alice seq = aliceSeq+3
+		require.Equal(t, aliceSeq+3, env.Seq(alice),
+			"alice seq should have advanced by 3 (outer + 2 inner payments)")
+
+		jtx.RequireBalance(t, env, alice, preAlice-uint64(jtx.XRP(3))-batchFee)
+		jtx.RequireBalance(t, env, bob, preBob+uint64(jtx.XRP(3)))
+	})
+
+	t.Run("outer batch terPRE_SEQ", func(t *testing.T) {
+		// Reference: rippled testSequenceOpenLedger() "Outer Batch terPRE_SEQ"
+		// Batch outer has a future sequence (carolSeq+1) -> terPRE_SEQ.
+		// A noop advances carol's seq. After close: batch succeeds.
+		env := newBatchEnv(t)
+		env.EnableOpenLedgerReplay()
+
+		alice := jtx.NewAccount("alice")
+		bob := jtx.NewAccount("bob")
+		carol := jtx.NewAccount("carol")
+		env.Fund(alice, bob, carol)
+		env.Close()
+
+		aliceSeq := env.Seq(alice)
+		carolSeq := env.Seq(carol)
+		preAlice := env.Balance(alice)
+		preBob := env.Balance(bob)
+
+		// Batch Txn with future carolSeq -> terPRE_SEQ
+		batchFee := CalcBatchFeeFromEnv(env, 1, 2)
+		batch := NewBatchBuilder(carol, carolSeq+1, batchFee, batchtx.BatchFlagAllOrNothing).
+			AddInnerTx(MakeInnerPaymentXRP(alice, bob, 1, aliceSeq)).
+			AddInnerTx(MakeInnerPaymentXRP(alice, bob, 2, aliceSeq+1)).
+			AddSigner(alice).
+			MustBuild()
+		result := env.Submit(batch)
+		jtx.RequireTxFail(t, result, "terPRE_SEQ")
+
+		// AccountSet noop at carolSeq -> tesSUCCESS (advances carol's seq)
+		noopTxn := accounttx.NewAccountSet(carol.Address)
+		noopTxn.Fee = fmt.Sprintf("%d", env.BaseFee())
+		noopTxn.SigningPubKey = carol.PublicKeyHex()
+		noopTxn.Sequence = &carolSeq
+		result = env.Submit(noopTxn)
+		jtx.RequireTxSuccess(t, result)
+		env.Close()
+
+		// After close: noop advances carol's seq, then batch succeeds.
+		// Close again to flush held transactions.
+		env.Close()
+
+		// Final state:
+		// - alice consumed aliceSeq and aliceSeq+1 (inner payments)
+		require.Equal(t, aliceSeq+2, env.Seq(alice),
+			"alice seq should advance by 2 (inner payments)")
+
+		// - carol consumed carolSeq (noop) and carolSeq+1 (batch outer)
+		require.Equal(t, carolSeq+2, env.Seq(carol),
+			"carol seq should advance by 2 (noop + batch outer)")
+
+		// alice paid XRP(3) total to bob
+		jtx.RequireBalance(t, env, alice, preAlice-uint64(jtx.XRP(3)))
+		jtx.RequireBalance(t, env, bob, preBob+uint64(jtx.XRP(3)))
+	})
+}
+
+func TestObjectsOpenLedger(t *testing.T) {
+	// Reference: rippled Batch_test.cpp testObjectsOpenLedger()
+	// Tests interactions between batch inner transactions creating/consuming
+	// ledger objects and standalone transactions that depend on those objects.
+
+	t.Run("consume object before batch txn", func(t *testing.T) {
+		// Reference: rippled testObjectsOpenLedger() "Consume Object Before Batch Txn"
+		// CheckCash submitted before the batch that creates the check.
+		// In rippled, CheckCash gets tecNO_ENTRY initially (inner txns deferred).
+		// During consensus replay, batch (alice) is applied first in canonical
+		// order, creating the check. Then CheckCash (bob) succeeds.
+		//
+		// The open ledger cannot see the Check created by the Batch inner until
+		// consensus replay closes the ledger.
+		env := newBatchEnv(t)
+		env.EnableOpenLedgerReplay()
+
+		alice := jtx.NewAccount("alice")
+		bob := jtx.NewAccount("bob")
+		env.Fund(alice, bob)
+		env.Close()
+
+		aliceTicketSeq := env.CreateTickets(alice, 10)
+		env.Close()
+
+		aliceSeq := env.Seq(alice)
+		bobSeq := env.Seq(bob)
+		preAlice := env.Balance(alice)
+		preBob := env.Balance(bob)
+
+		// CheckCash Txn for a check that doesn't exist yet
+		chkID := GetCheckIndex(alice, aliceSeq)
+		cashTxn := check.NewCheckCash(bob.Address, chkID)
+		cashTxn.SetExactAmount(tx.NewXRPAmount(jtx.XRP(10)))
+		cashTxn.Fee = fmt.Sprintf("%d", env.BaseFee())
+		cashTxn.SigningPubKey = bob.PublicKeyHex()
+		cashTxn.Sequence = &bobSeq
+		cashResult := env.Submit(cashTxn)
+		jtx.RequireTxClaimed(t, cashResult, "tecNO_ENTRY")
+
+		// Batch Txn: creates the check and pays XRP(1)
+		batchFee := CalcBatchFeeFromEnv(env, 0, 2)
+		batch := NewBatchBuilderWithTicket(alice, aliceTicketSeq, batchFee, batchtx.BatchFlagAllOrNothing).
+			AddInnerTx(MakeInnerCheckCreate(alice, bob, tx.NewXRPAmount(jtx.XRP(10)), aliceSeq)).
+			AddInnerTx(MakeInnerPaymentXRPWithTicket(alice, bob, 1, aliceTicketSeq+1)).
+			MustBuild()
+		result := env.Submit(batch)
+		jtx.RequireTxSuccess(t, result)
+		env.Close()
+
+		// After close (replay): batch creates check, then CheckCash succeeds.
+		env.Close()
+
+		// Final state verification:
+		// bob should have cashed XRP(10) check + received XRP(1) payment
+		// alice should have lost XRP(10) (check) + XRP(1) (payment) + batchFee + baseFee (for CheckCash on bob)
+		// bob consumes bobSeq for CheckCash
+		require.Equal(t, bobSeq+1, env.Seq(bob),
+			"bob seq should advance by 1 (CheckCash)")
+		// alice consumes aliceSeq (inner CheckCreate), aliceTicketSeq (outer), aliceTicketSeq+1 (inner payment)
+		require.Equal(t, aliceSeq+1, env.Seq(alice),
+			"alice seq should advance by 1 (inner CheckCreate)")
+
+		// Balance check: alice paid XRP(10) check + XRP(1) + batchFee
+		jtx.RequireBalance(t, env, alice, preAlice-uint64(jtx.XRP(11))-batchFee)
+		// bob gained XRP(10) check + XRP(1), paid baseFee for CheckCash
+		jtx.RequireBalance(t, env, bob, preBob+uint64(jtx.XRP(11))-env.BaseFee())
+	})
+
+	t.Run("create object before batch txn", func(t *testing.T) {
+		// Reference: rippled testObjectsOpenLedger() "Create Object Before Batch Txn"
+		// CheckCreate submitted before the batch. The batch's inner CheckCash
+		// consumes the check. The standalone CheckCreate runs first in the
+		// open view.
+		env := newBatchEnv(t)
+		env.EnableOpenLedgerReplay()
+
+		alice := jtx.NewAccount("alice")
+		bob := jtx.NewAccount("bob")
+		env.Fund(alice, bob)
+		env.Close()
+
+		aliceTicketSeq := env.CreateTickets(alice, 10)
+		env.Close()
+
+		aliceSeq := env.Seq(alice)
+		bobSeq := env.Seq(bob)
+		preAlice := env.Balance(alice)
+		preBob := env.Balance(bob)
+
+		// CheckCreate Txn (standalone) — alice creates a check payable to bob
+		chkID := GetCheckIndex(alice, aliceSeq)
+		createTxn := check.NewCheckCreate(alice.Address, bob.Address, tx.NewXRPAmount(jtx.XRP(10)))
+		createTxn.Fee = fmt.Sprintf("%d", env.BaseFee())
+		createTxn.SigningPubKey = alice.PublicKeyHex()
+		createTxn.Sequence = &aliceSeq
+		result := env.Submit(createTxn)
+		jtx.RequireTxSuccess(t, result)
+
+		// Batch Txn: inner CheckCash (bob cashes the check) + inner Payment
+		batchFee := CalcBatchFeeFromEnv(env, 1, 2)
+		batch := NewBatchBuilderWithTicket(alice, aliceTicketSeq, batchFee, batchtx.BatchFlagAllOrNothing).
+			AddInnerTx(MakeInnerCheckCash(bob, chkID, tx.NewXRPAmount(jtx.XRP(10)), bobSeq)).
+			AddInnerTx(MakeInnerPaymentXRPWithTicket(alice, bob, 1, aliceTicketSeq+1)).
+			AddSigner(bob).
+			MustBuild()
+		result = env.Submit(batch)
+		jtx.RequireTxSuccess(t, result)
+		env.Close()
+
+		// Final state verification:
+		// bob cashed check XRP(10) + received XRP(1) payment
+		require.Equal(t, bobSeq+1, env.Seq(bob),
+			"bob seq should advance by 1 (inner CheckCash)")
+		require.Equal(t, aliceSeq+1, env.Seq(alice),
+			"alice seq should advance by 1 (standalone CheckCreate)")
+
+		// alice paid: XRP(10) check + XRP(1) payment + batchFee + baseFee for CheckCreate
+		jtx.RequireBalance(t, env, alice, preAlice-uint64(jtx.XRP(11))-batchFee-env.BaseFee())
+		// bob gained: XRP(10) check + XRP(1) payment
+		jtx.RequireBalance(t, env, bob, preBob+uint64(jtx.XRP(11)))
+	})
+
+	t.Run("after batch txn", func(t *testing.T) {
+		// Reference: rippled testObjectsOpenLedger() "After Batch Txn"
+		// Batch creates a check (inner), then standalone CheckCash tries to cash it.
+		// In rippled, the CheckCash gets tecNO_ENTRY because batch inner txns
+		// are deferred. During replay, batch applies first, then CheckCash succeeds.
+		env := newBatchEnv(t)
+		env.EnableOpenLedgerReplay()
+
+		alice := jtx.NewAccount("alice")
+		bob := jtx.NewAccount("bob")
+		env.Fund(alice, bob)
+		env.Close()
+
+		aliceTicketSeq := env.CreateTickets(alice, 10)
+		env.Close()
+
+		aliceSeq := env.Seq(alice)
+		bobSeq := env.Seq(bob)
+		preAlice := env.Balance(alice)
+		preBob := env.Balance(bob)
+
+		// Batch Txn: creates check + payment
+		batchFee := CalcBatchFeeFromEnv(env, 0, 2)
+		chkID := GetCheckIndex(alice, aliceSeq)
+		batch := NewBatchBuilderWithTicket(alice, aliceTicketSeq, batchFee, batchtx.BatchFlagAllOrNothing).
+			AddInnerTx(MakeInnerCheckCreate(alice, bob, tx.NewXRPAmount(jtx.XRP(10)), aliceSeq)).
+			AddInnerTx(MakeInnerPaymentXRPWithTicket(alice, bob, 1, aliceTicketSeq+1)).
+			MustBuild()
+		result := env.Submit(batch)
+		jtx.RequireTxSuccess(t, result)
+
+		// CheckCash Txn (standalone) — the open view has not executed Batch inners.
+		cashTxn := check.NewCheckCash(bob.Address, chkID)
+		cashTxn.SetExactAmount(tx.NewXRPAmount(jtx.XRP(10)))
+		cashTxn.Fee = fmt.Sprintf("%d", env.BaseFee())
+		cashTxn.SigningPubKey = bob.PublicKeyHex()
+		cashTxn.Sequence = &bobSeq
+		cashResult := env.Submit(cashTxn)
+		jtx.RequireTxClaimed(t, cashResult, "tecNO_ENTRY")
+		env.Close()
+
+		// After close (replay): batch creates check, then CheckCash succeeds.
+		env.Close()
+
+		// Final state verification:
+		require.Equal(t, bobSeq+1, env.Seq(bob),
+			"bob seq should advance by 1 (CheckCash)")
+		require.Equal(t, aliceSeq+1, env.Seq(alice),
+			"alice seq should advance by 1 (inner CheckCreate)")
+
+		// alice lost XRP(10) check + XRP(1) payment + batchFee
+		jtx.RequireBalance(t, env, alice, preAlice-uint64(jtx.XRP(11))-batchFee)
+		// bob gained XRP(10) check + XRP(1) payment, paid baseFee for CheckCash
+		jtx.RequireBalance(t, env, bob, preBob+uint64(jtx.XRP(11))-env.BaseFee())
+	})
+}
+
+func TestOpenLedger(t *testing.T) {
+	// Reference: rippled Batch_test.cpp testOpenLedger()
+	// Tests a mixed scenario: alice pays bob, then an atomic batch with
+	// alice+bob, then bob pays alice with a future sequence (terPRE_SEQ).
+	// The canonical ordering during consensus determines transaction placement.
+	//
+	// In rippled's canonical ordering (salted), alice's payment comes first,
+	// then the batch, then bob's payment is retried next ledger.
+	//
+	// The open ledger cannot see the sequence consumed by the Batch inner until
+	// consensus replay closes the ledger.
+	env := newBatchEnv(t)
+	env.EnableOpenLedgerReplay()
+
+	alice := jtx.NewAccount("alice")
+	bob := jtx.NewAccount("bob")
+	env.Fund(alice, bob)
+	env.Close()
+
+	// Extra noop to advance bob's seq (matching rippled: env(noop(bob)))
+	noopBob := accounttx.NewAccountSet(bob.Address)
+	noopBob.Fee = fmt.Sprintf("%d", env.BaseFee())
+	noopBob.SigningPubKey = bob.PublicKeyHex()
+	bobNoopSeq := env.Seq(bob)
+	noopBob.Sequence = &bobNoopSeq
+	result := env.Submit(noopBob)
+	jtx.RequireTxSuccess(t, result)
+	env.Close()
+
+	aliceSeq := env.Seq(alice)
+	preAlice := env.Balance(alice)
+	preBob := env.Balance(bob)
+	bobSeq := env.Seq(bob)
+
+	// Alice Pays Bob (Open Ledger)
+	payTxn1 := payment.NewPayment(alice.Address, bob.Address, tx.NewXRPAmount(jtx.XRP(10)))
+	payTxn1.Fee = fmt.Sprintf("%d", env.BaseFee())
+	payTxn1.SigningPubKey = alice.PublicKeyHex()
+	payTxn1.Sequence = &aliceSeq
+	result = env.Submit(payTxn1)
+	jtx.RequireTxSuccess(t, result)
+
+	// Alice & Bob Atomic Batch
+	batchFee := CalcBatchFeeFromEnv(env, 1, 2)
+	batch := NewBatchBuilder(alice, aliceSeq+1, batchFee, batchtx.BatchFlagAllOrNothing).
+		AddInnerTx(MakeInnerPaymentXRP(alice, bob, 10, aliceSeq+2)).
+		AddInnerTx(MakeInnerPaymentXRP(bob, alice, 5, bobSeq)).
+		AddSigner(bob).
+		MustBuild()
+	result = env.Submit(batch)
+	jtx.RequireTxSuccess(t, result)
+
+	// Bob pays Alice (Open Ledger) at bobSeq+1. The Batch inner that consumes
+	// bobSeq is deferred until close, so this transaction is held for retry.
+	bobPaySeq := bobSeq + 1
+	payTxn2 := payment.NewPayment(bob.Address, alice.Address, tx.NewXRPAmount(jtx.XRP(5)))
+	payTxn2.Fee = fmt.Sprintf("%d", env.BaseFee())
+	payTxn2.SigningPubKey = bob.PublicKeyHex()
+	payTxn2.Sequence = &bobPaySeq
+	payResult2 := env.Submit(payTxn2)
+	jtx.RequireTxFail(t, payResult2, "terPRE_SEQ")
+	env.Close()
+
+	// Close again to ensure any held transactions are retried
+	env.Close()
+
+	// Final state verification (matches rippled):
+	// alice: aliceSeq (pay#1) + aliceSeq+1 (batch outer) + aliceSeq+2 (inner pay) = 3 txns
+	require.Equal(t, aliceSeq+3, env.Seq(alice),
+		"alice seq should have advanced by 3")
+
+	// bob: bobSeq (inner pay to alice) + bobSeq+1 (standalone pay to alice) = 2 txns
+	require.Equal(t, bobSeq+2, env.Seq(bob),
+		"bob seq should have advanced by 2")
+
+	// Balance verification:
+	// alice: -XRP(10) pay1 - baseFee pay1 - XRP(10) inner pay - batchFee + XRP(5) inner bob->alice
+	// but also bob pays alice XRP(5) standalone
+	// Net alice: preAlice - XRP(10) - XRP(10) + XRP(5) + XRP(5) - baseFee - batchFee
+	//          = preAlice - XRP(10) - baseFee - batchFee
+	jtx.RequireBalance(t, env, alice, preAlice-uint64(jtx.XRP(10))-batchFee-env.BaseFee())
+
+	// bob: +XRP(10) pay1 + XRP(10) inner alice->bob - XRP(5) inner bob->alice
+	//      - XRP(5) standalone - baseFee for standalone
+	// Net bob: preBob + XRP(10) - baseFee
+	jtx.RequireBalance(t, env, bob, preBob+uint64(jtx.XRP(10))-env.BaseFee())
+}
+
+func TestBadRawTxn(t *testing.T) {
+	t.Run("nil inner transaction", func(t *testing.T) {
+		env := newBatchEnv(t)
+		alice := jtx.NewAccount("alice")
+		bob := jtx.NewAccount("bob")
+		env.Fund(alice, bob)
+		env.Close()
+
+		seq := env.Seq(alice)
+		batchFee := CalcBatchFeeFromEnv(env, 0, 2)
+
+		// Manually build a batch with a nil inner tx
+		batch := batchtx.NewBatch(alice.Address)
+		batch.Fee = fmt.Sprintf("%d", batchFee)
+		batch.SetSequence(seq)
+		batch.SetFlags(batchtx.BatchFlagAllOrNothing)
+		batch.RawTransactions = []batchtx.RawTransaction{
+			{RawTransaction: batchtx.RawTransactionData{InnerTx: nil}}, // nil
+			{RawTransaction: batchtx.RawTransactionData{InnerTx: MakeInnerPaymentXRP(alice, bob, 1, seq+2)}},
+		}
+
+		result := env.Submit(batch)
+		jtx.RequireTxFail(t, result, "temMALFORMED")
+		jtx.RequireSequence(t, env, alice, seq)
+	})
+}
